// color-generator/models/types/ThemeNumberOfColors.ts
export enum ThemeNumberOfColors {
    SIMPLE = 'simple',
    BI_COLOR_UP = 'bi-colors-up',
    TRI_COLOR_UP = 'tri-colors-up',
    BI_COLOR_DOWN = 'bi-colors-down',
    TRI_COLOR_DOWN = 'tri-colors-down',
    BI_COLOR_CENTERED = 'bi-colors-centered',
    TRI_COLOR_CENTERED = 'tri-colors-centered',
}